scripts/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
PyYAML>=6.0
